move sub-level chain-event schemas + add getRoutingKey function/types
