+++ libs/core/src/integration/index.ts
@@ -1,3 +1,4 @@
+export * from './chain-event.schemas';
 export * from './email.schemas';
 export * from './events';
 export * from './events.schemas';
