+++ libs/core/src/ports/in-memory-brokers.ts
@@ -1,9 +1,12 @@
+import { RoutingKey } from '@hicommonwealth/core';
+
 export const successfulInMemoryBroker = {
   name: 'successful-in-memory-broker',
   dispose: () => Promise.resolve(),
   publish: () => Promise.resolve(true),
   subscribe: () => Promise.resolve(true),
   isHealthy: () => Promise.resolve(true),
+  getRoutingKey: (): RoutingKey => 'ThreadCreated',
 };
 
 export const failingInMemoryBroker = {
