+++ packages/commonwealth/server/workers/commonwealthConsumer/policies/chainEventCreated/handleCommunityStakeTrades.ts
@@ -1,4 +1,4 @@
-import { events, logger } from '@hicommonwealth/core';
+import { CommunityStakeTrade, events, logger } from '@hicommonwealth/core';
 import { DB } from '@hicommonwealth/model';
 import { BigNumber } from 'ethers';
 import { fileURLToPath } from 'url';
@@ -20,7 +20,7 @@ export async function handleCommunityStakeTrades(
     4: ethAmount,
     // 5: protocolEthAmount,
     // 6: nameSpaceEthAmount,
-  } = event.parsedArgs as z.infer<typeof events.CommunityStakeTrade>;
+  } = event.parsedArgs as z.infer<typeof CommunityStakeTrade>;
 
   const existingTxn = await models.StakeTransaction.findOne({
     where: {
