+++ packages/commonwealth/server/workers/knock/eventHandlers/chainEvents/handleCommunityStakeTrades.ts
@@ -1,4 +1,5 @@
 import {
+  CommunityStakeTrade,
   events,
   logger,
   notificationsProvider,
@@ -18,7 +19,7 @@ export async function handleCommunityStakeTrades(
   event: z.infer<typeof events.ChainEventCreated>,
 ) {
   const { 1: namespaceAddress, 2: isBuy } = event.parsedArgs as z.infer<
-    typeof events.CommunityStakeTrade
+    typeof CommunityStakeTrade
   >;
 
   const community = await models.Community.findOne({
@@ -37,11 +38,11 @@ export async function handleCommunityStakeTrades(
 
   const users = await models.sequelize.query<{ id: string }>(
     `
-        SELECT DISTINCT(user_id)::TEXT as id
-        FROM "Addresses"
-        WHERE community_id = :communityId
-          AND role = 'admin';
-    `,
+            SELECT DISTINCT(user_id)::TEXT as id
+            FROM "Addresses"
+            WHERE community_id = :communityId
+              AND role = 'admin';
+        `,
     {
       raw: true,
       type: QueryTypes.SELECT,
